+++ WMI/WMI/DataProviders/PagingFileDataProvider.cs
@@ -8,7 +8,7 @@ namespace WMI.DataProviders
 			: base(updateInterval)
 		{
 			AddSearcher("Win32_PerfFormattedData_PerfOS_PagingFile",
-				new PropertySettersDictionary<PagingFile>() {{"PercentUsage", (file, o) => file.UsagePercent = (byte) o}},
+				new PropertySettersDictionary<PagingFile>() {{"PercentUsage", (file, o) => file.UsagePercent = (uint) o}},
 				"Name = '_Total'");
 		}
 	}
